requirements.txt
Flask==2.3.2
PyYAML==6.0.1
requests==2.31.0
urllib3==2.0.4
